+++ src/Sentry/Protocol/DeviceOrientation.cs
@@ -1,6 +1,5 @@
 using System.Runtime.Serialization;
 
-// ReSharper disable once CheckNamespace
 namespace Sentry.Protocol
 {
     /// <summary>